Linted the rest, added stylelint
