+++ .eslintrc.json
@@ -2,9 +2,35 @@
   "extends": [
     "airbnb"
   ],
+  "env": {
+    "browser": true,
+    "node": true,
+    "jasmine": true
+  },
   "rules": {
     "consistent-return": "off",
     "no-plusplus": "off",
-    "object-curly-newline": "off"
+    "object-curly-newline": "off",
+    "react/react-in-jsx-scope": "off",
+    "react/prop-types": "off",
+    "no-underscore-dangle": "off",
+    "react/jsx-props-no-spreading": "off",
+    "react/no-unescaped-entities": "off",
+    "func-names": "off",
+    "react/jsx-one-expression-per-line": "off",
+    "jsx-a11y/anchor-is-valid": "off",
+    "react/no-array-index-key": "off",
+    "prefer-destructuring": "off",
+    "jsx-a11y/click-events-have-key-events": "off",
+    "jsx-a11y/no-noninteractive-element-interactions": "off",
+    "require-jsdoc": ["error", {
+      "require": {
+          "FunctionDeclaration": true,
+          "MethodDefinition": false,
+          "ClassDeclaration": true,
+          "ArrowFunctionExpression": true,
+          "FunctionExpression": false
+      }
+  }]
   }
 }
