+++ package.json
@@ -48,6 +48,7 @@
         "react-icons": "^3.10.0",
         "react-transition-group": "^4.4.1",
         "styled-components": "^5.1.1",
+        "stylelint-config-sass-guidelines": "^7.0.0",
         "uuid": "^8.2.0"
     },
     "devDependencies": {
@@ -59,6 +60,10 @@
         "eslint-plugin-jsx-a11y": "^6.3.1",
         "eslint-plugin-prettier": "^3.1.4",
         "eslint-plugin-react": "^7.20.3",
-        "eslint-plugin-react-hooks": "^4.0.0"
+        "eslint-plugin-react-hooks": "^4.0.0",
+        "stylelint": "^13.6.1",
+        "stylelint-config-idiomatic-order": "^8.1.0",
+        "stylelint-config-standard": "^20.0.0",
+        "stylelint-order": "^4.1.0"
     }
 }
