--- src/pages/_document.js
@@ -1,58 +1,0 @@
-/**
- * This overwrites Next.js's default document to allow styled-components to be inline
- * with Next's server side rendering and also set default meta tags. 
- */
-
-import { ServerStyleSheet as StyledComponentSheets } from "styled-components";
-import { ServerStyleSheets as MaterialUiServerStyleSheets } from "@material-ui/styles";
-import Document, { Html, Head, Main, NextScript } from "next/document";
-
-export default class MyDocument extends Document {
-    static async getInitialProps(ctx) {
-        const styledComponentSheet = new StyledComponentSheets();
-        const materialUiSheets = new MaterialUiServerStyleSheets();
-        const originalRenderPage = ctx.renderPage;
-        try {
-            ctx.renderPage = () =>
-                originalRenderPage({
-                    enhanceApp: App => props =>
-                        styledComponentSheet.collectStyles(
-                            materialUiSheets.collect(<App {...props} />)
-                        ),
-                });
-            const initialProps = await Document.getInitialProps(ctx);
-            return {
-                ...initialProps,
-                styles: [
-                    <React.Fragment key="styles">
-                        {initialProps.styles}
-                        {materialUiSheets.getStyleElement()}
-                        {styledComponentSheet.getStyleElement()}
-                    </React.Fragment>,
-                ],
-            };
-        } finally {
-            styledComponentSheet.seal();
-        }
-    }
-
-    render() {
-        return (
-            <Html>
-                <Head>
-                    <meta name="viewport" content="width=device-width, initial-scale=1" />
-                    <meta charSet="utf-8" />
-                    <link
-                        rel="shortcut icon"
-                        type="image/x-icon"
-                        href="/favicon.ico"
-                    />
-                </Head>
-                <body>
-                    <Main />
-                    <NextScript />
-                </body>
-            </Html>
-        );
-    }
-}
