+++ src/pages/_document.jsx
@@ -1,0 +1,56 @@
+import { Fragment } from 'react';
+import { ServerStyleSheet as StyledComponentSheets } from 'styled-components';
+import { ServerStyleSheets as MaterialUiServerStyleSheets } from '@material-ui/styles';
+import Document, { Html, Head, Main, NextScript } from 'next/document';
+
+/**
+ * This overwrites Next.js default document to allow styled-components to be inline
+ * with Next's server side rendering and also set default meta tags.
+ */
+export default class MyDocument extends Document {
+  static async getInitialProps(ctx) {
+    const styledComponentSheet = new StyledComponentSheets();
+    const materialUiSheets = new MaterialUiServerStyleSheets();
+    const originalRenderPage = ctx.renderPage;
+    try {
+      ctx.renderPage = () => originalRenderPage({
+        enhanceApp: (App) => (props) => styledComponentSheet.collectStyles(
+          materialUiSheets.collect(<App {...props} />),
+        ),
+      });
+      const initialProps = await Document.getInitialProps(ctx);
+      return {
+        ...initialProps,
+        styles: [
+          <Fragment key="styles">
+            {initialProps.styles}
+            {materialUiSheets.getStyleElement()}
+            {styledComponentSheet.getStyleElement()}
+          </Fragment>,
+        ],
+      };
+    } finally {
+      styledComponentSheet.seal();
+    }
+  }
+
+  render() {
+    return (
+      <Html>
+        <Head>
+          <meta name="viewport" content="width=device-width, initial-scale=1" />
+          <meta charSet="utf-8" />
+          <link
+            rel="shortcut icon"
+            type="image/x-icon"
+            href="/favicon.ico"
+          />
+        </Head>
+        <body>
+          <Main />
+          <NextScript />
+        </body>
+      </Html>
+    );
+  }
+}
