+++ src/pages/finances.jsx
@@ -1,18 +1,21 @@
-import styled from "styled-components";
+import styled from 'styled-components';
 
 const StyledFinances = styled.div`
+    left: 50vw;
     position: absolute;
     top: 50vh;
-    left: 50vw;
     transform: translate(-50%, -50%);
 `;
 
+/**
+ * TODO
+ */
 function Finances() {
-    return (
-        <StyledFinances>
-            <p>Work in progress...</p>
-        </StyledFinances>
-    );
+  return (
+    <StyledFinances>
+      <p>Work in progress...</p>
+    </StyledFinances>
+  );
 }
 
 export default Finances;
